Need to Cast to INT

+++ submit_jobs-HPC.py
@@ -36,14 +36,14 @@ for i in listArray:
 		tempSplit = i.split("/")
 		if "-" in tempSplit[1]:
 			tempSplit = tempSplit[1].split("-")
-			coresAvailable["node17"] = coresAvailable["node17"] - (tempSplit[1] - tempSplit[0] + 1)
+			coresAvailable["node17"] = coresAvailable["node17"] - (int)(tempSplit[1] - tempSplit[0] + 1)
 		else:
 			coresAvailable["node17"] = coresAvailable["node17"] - 1
 	if "n018" in i:
 		tempSplit = i.split("/")
 		if "-" in tempSplit[1]:
 			tempSplit = tempSplit[1].split("-")
-			coresAvailable["node18"] = coresAvailable["node18"] - (tempSplit[1] - tempSplit[0] + 1)
+			coresAvailable["node18"] = coresAvailable["node18"] - (int)(tempSplit[1] - tempSplit[0] + 1)
 		else:
 			coresAvailable["node18"] = coresAvailable["node18"] - 1
